backend/requirements.txt
Flask==2.2.3
gunicorn==20.1.0
supabase==1.0.3
python-dotenv==1.0.0
requests==2.28.1
Flask-CORS==3.0.10
google-cloud-speech==2.21.0
scikit-learn==1.2.1
tensorflow==2.9.1
protobuf==3.19.6
numpy==1.24.2
pandas==1.5.3
joblib==1.1.1
werkzeug==2.2.3
pyaudio==0.2.13
SQLAlchemy==2.0.12
psycopg2-binary==2.9.6
redis==4.5.1
selenium==4.15.0
webdriver-manager==4.0.1